api/requirements.txt
azure-functions
azure-storage-blob==12.19.0
PyJWT==2.8.0
pandas==2.1.0
openpyxl==3.1.2
python-dateutil==2.8.2
playwright==1.40.0
